+++ thinqconnect/integration/__init__.py
@@ -8,6 +8,7 @@
 from .homeassistant.api import (
     HABridge,
     async_get_ha_bridge_list,
+    NotConnectedDeviceError,
 )
 from .homeassistant.property import ActiveMode
 from .homeassistant.specification import (
@@ -15,12 +16,14 @@ from .homeassistant.specification import (
     ThinQPropertyEx,
     TimerProperty,
 )
-from .homeassistant.state import PropertyState
+from .homeassistant.state import DeviceState, PropertyState
 
 __all__ = [
     "ActiveMode",
+    "DeviceState",
     "ExtendedProperty",
     "HABridge",
+    "NotConnectedDeviceError",
     "PropertyState",
     "ThinQPropertyEx",
     "TimerProperty",