+++ RELEASE_NOTES.md
@@ -1,9 +1,15 @@
 # Release Notes
 
-## 1.0.4 (2025-02-)
+## 1.0.5 (2025-03-14)
+### Features
+* Add ventilator device
+* Support **current job mode** property control (dehumidifier)
+
+## 1.0.4 (2025-02-03)
 ### Features
 *  Add **cycle count** property (washer): [#7](https://github.com/thinq-connect/pythinqconnect/issues/7)
-*  Add **hop oil** and **flaver info** property for two capsules (home brew)
+*  Add **express mode name**, **express fridge** property (refrigerator)
+*  Add **hop oil** and **flavor info** property for two capsules (home_brew)
 *  Add **room temp mode**, **room water mode** property (system_boiler)
 *  Add **two set enabled** property (air_conditioner)
 *  Add **display light** property (air_conditioner): [#2](https://github.com/thinq-connect/pythinqconnect/issues/2)
@@ -11,7 +17,7 @@
 ### Deprecations
 *  Remove **remain time** properties (hood)
 ### Improvements
-*  Add temperature properties in Fahrenheit (air_conditioner, system_boiler, refrigerator, wine_cellar)
+*  Add temperature properties in Fahrenheit (air_conditioner, system_boiler, refrigerator, wine_cellar, water_heater)
 *  Add **temperature unit** property (oven)
 ### Fixes
 *  Fix **target temperature** property value from dict to number (oven)